Split gol.h/.cpp into Cell.h/.cpp and Board.h/.cpp
