+++ CMakeLists.txt
@@ -6,6 +6,6 @@ set(CMAKE_CXX_STANDARD 14)
 include_directories(.)
 
 add_executable(game_of_life
-        gol.cpp
-        gol.h
-        main.cpp)
+        Board.cpp
+        Board.h
+        main.cpp Cell.cpp Cell.h)
