+++ Cell.cpp
@@ -1,0 +1,12 @@
+#include "Cell.h"
+
+Cell::Cell() {
+    neighbors = 0;
+    state = false;
+}
+
+bool Cell::isAlive() const {
+    if(state)
+        return true;
+    return false;
+}
